♻️refactor: 리뷰 삭제 (@AuthenticationPrincipal 적용)

+++ src/main/java/com/project/likelion13thbe/domain/review/service/command/ReviewCommandServiceImpl.java
@@ -57,9 +57,13 @@ public class ReviewCommandServiceImpl implements ReviewCommandService {
     }
 
     @Override
-    public void deleteReview(Long reviewId) {
+    public void deleteReview(Long reviewId, Member member) {
         Review review = reviewRepository.findByIdAndNotDeleted(reviewId)
                 .orElseThrow(() -> new ReviewException(ReviewErrorCode.REVIEW_NOT_FOUND));
+
+        if (!review.getMember().getId().equals(member.getId())) {
+            throw new ReviewException(ReviewErrorCode.UNAUTHORIZED);
+        }
 
         review.delete();
     }
